migrate test_devinfo to pytest
